add tests (and move json specific tests)
